Fix fields not being set at all.
This removes the request header `X-Response-Control` as it was set to `minified` which caused the API to send a different set of data back which did not decode back to the given structure properly.

In the future, there are plans to give the developer full control over the output by providing other variants of the struct.

Fixes #6.